Add multi-game daily challenges, Hall of Fame tabs, and Supabase migration 006
- Add challenge_id to daily completions (migration 006); per-game HoF and account stats
- Blind Rank daily: daily-seed.js, challenge_id blind_ranking, scoped admin reset
- New daily pages: Higher or Lower and Codebreaker (medium, seeded)
- Hub: Blind Rank Daily Challenge name, two new daily cards, dailyChallenge rail routing
- Hall of Fame tabs; getHallOfFame(limit, challengeId); My Stats daily_challenges blocks
- Fix COALESCE int/text mix in game_rollup (004 and 006)

Made-with: Cursor

+++ supabase/migrations/004_account_stats_and_hall_of_fame.sql
@@ -103,11 +103,11 @@ mimi AS (
 ),
 game_rollup AS (
     SELECT
-        COALESCE(SUM(COALESCE((gs.stats->'playerStats'->>'gamesPlayed')::int, gs.stats->>'gamesPlayed', '0')::int), 0)::int AS total_games_played,
+        COALESCE(SUM((COALESCE(gs.stats->'playerStats'->>'gamesPlayed', gs.stats->>'gamesPlayed', '0'))::int), 0)::int AS total_games_played,
         COALESCE(
             jsonb_object_agg(
                 gs.game_id,
-                COALESCE((gs.stats->'playerStats'->>'gamesPlayed')::int, gs.stats->>'gamesPlayed', '0')::int
+                (COALESCE(gs.stats->'playerStats'->>'gamesPlayed', gs.stats->>'gamesPlayed', '0'))::int
             ),
             '{}'::jsonb
         ) AS games_played_by_game
